Updated footer in page

+++ app/Http/Controllers/MissionRatingController.php
@@ -2,18 +2,28 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\MissionApplication;
 use App\Models\MissionRating;
+use Illuminate\Http\Client\Response;
 use Illuminate\Http\Request;
 
 class MissionRatingController extends Controller
 {
     public function addRating(Request $request){
-        $rate = MissionRating::where('user_id','=',$request->user_id)
+        $checkUser = MissionApplication::where('user_id',$request->user_id)
+                                        ->where('mission_id',$request->mission_id)
+                                        ->first();
+        if($checkUser!=Null && $checkUser->approval_status=="APPROVE"){
+            $rate = MissionRating::where('user_id','=',$request->user_id)
                             ->where('mission_id','=',$request->mission_id);
-        if($rate!=null){
-            $rate->delete();
+            if($rate!=null){
+                $rate->delete();
+            }
+            MissionRating::create($request->post());
+            return('Successfully added');
         }
-        MissionRating::create($request->post());
-        return('Successfully added');
+        else{
+            return "Sorry you can't Add rating";
+        }
     }
 }
